+++ CHANGES.rst
@@ -4,9 +4,10 @@ Changes
 0.5.0 (unreleased)
 ==================
 
-- Connections have a new ``read_only_cursor`` method for obtaining a
-  `database cursor <http://initd.org/psycopg/docs/cursor.html>`_ for
-  selecting data and for using the `cursor's mogrify method
+- The ``newt.db.search`` module has a new ``read_only_cursor``
+  function for obtaining a `database cursor
+  <http://initd.org/psycopg/docs/cursor.html>`_ for selecting data and
+  for using the `cursor's mogrify method
   <http://initd.org/psycopg/docs/cursor.html#cursor.mogrify>`_.
 
 - The helpers for setting up full-text search indexes now accept a